Combines plan/interest usecases to one, so we do not have separate files of interests<->plans any longer, one file per interest/plan-combination
