+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-cycling.js
@@ -3,9 +3,9 @@ import {
   vicinityScoreSubQuery,
   sparqlQuery,
 } from "../../app/sparql-builder-utils.js";
+import won from "../../app/service/won.js";
 import { getIn } from "../../app/utils.js";
 import { Generator } from "sparqljs";
-import won from "../../app/service/won.js";
 import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
 
 window.SparqlGenerator4dbg = Generator;
@@ -76,8 +76,87 @@ export const cyclingPlan = {
         `?seeks s:object ?planObject.`,
         `?planObject s:about <http://dbpedia.org/resource/Cycling>.`,
         `?thisAtom hold:heldBy/buddy:buddy/hold:holds ${resultName}.`,
-        `BIND( ( 
+        `BIND( (
           COALESCE(?location_geoScore, 0) 
+        ) / 5  as ?score)`,
+        // `FILTER(?score > 0)`, // not necessary atm to filter; there are parts of -postings we can't match yet (e.g. NLP on description). also content's sparse anyway.
+      ],
+      orderBy: [{ order: "DESC", variable: "?score" }],
+    });
+
+    return query;
+  },
+};
+
+export const cyclingInterest = {
+  identifier: "cyclingInterest",
+  label: "Add Cycling Interest",
+  icon: "#ico36_uc_cycling_cropped",
+  draft: {
+    ...mergeInEmptyDraft({
+      content: {
+        type: ["won:Interest"],
+        title: "I am interested in cycling!",
+      },
+      seeks: {
+        type: ["s:PlanAction"],
+        eventObject: "http://dbpedia.org/resource/Cycling",
+      },
+    }),
+  },
+  enabledUseCases: ["cyclingPlan"],
+  reactionUseCases: ["cyclingPlan"],
+  details: {
+    title: { ...details.title },
+    description: { ...details.description },
+    location: {
+      ...details.location,
+      mandatory: true,
+    },
+  },
+  seeksDetails: {},
+
+  generateQuery: (draft, resultName) => {
+    const vicinityScoreSQ = vicinityScoreSubQuery({
+      resultName: resultName,
+      bindScoreAs: "?location_geoScore",
+      pathToGeoCoords: "s:location/s:geo",
+      prefixesInPath: {
+        s: won.defaultContext["s"],
+        won: won.defaultContext["won"],
+      },
+      geoCoordinates: getIn(draft, ["content", "location"]),
+    });
+
+    const subQueries = [vicinityScoreSQ]
+      .filter(sq => sq) // filter out non-existing details (the SQs should be `undefined` for them)
+      .map(sq => ({
+        query: sq,
+        optional: true, // so counterparts without that detail don't get filtered out (just assigned a score of 0 via `coalesce`)
+      }));
+
+    const query = sparqlQuery({
+      prefixes: {
+        won: won.defaultContext["won"],
+        buddy: won.defaultContext["buddy"],
+        hold: won.defaultContext["hold"],
+        rdf: won.defaultContext["rdf"],
+        s: won.defaultContext["s"],
+      },
+      distinct: true,
+      variables: [resultName, "?score"],
+      subQueries: subQueries,
+      where: [
+        `${resultName} rdf:type won:Atom.`,
+        `${resultName} rdf:type s:PlanAction.`,
+        `${resultName} s:object ?planObject.`,
+        `${resultName} hold:heldBy ?holder.`,
+        `?thisAtom hold:heldBy/buddy:buddy/^hold:holds ${resultName}`,
+        `?planObject s:about <http://dbpedia.org/resource/Cycling>.`,
+        `?thisAtom hold:heldBy/buddy:buddy/hold:holds ${resultName}.`,
+        // calculate average of scores; can be weighed if necessary
+        `BIND( (
+          COALESCE(?location_geoScore, 0)
         ) / 5  as ?score)`,
         // `FILTER(?score > 0)`, // not necessary atm to filter; there are parts of -postings we can't match yet (e.g. NLP on description). also content's sparse anyway.
       ],
